+++ tests/conftest.py
@@ -42,7 +42,10 @@ def get_publisher_account():
 
 
 def get_consumer_account():
-    return get_account(0)
+    os.environ['PARITY_ADDRESS1'] = '0x068ed00cf0441e4829d9784fcbe7b9e26d4bd8d0'
+    os.environ['PARITY_PASSWORD1'] = 'secret'
+    os.environ['PARITY_KEYFILE1'] = 'tests/resources/data/consumer_key_file.json'
+    return get_account(1)
 
 
 def get_sample_ddo():