+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 History
 =======
 
+0.3.0 (May, 2020)
+-------------------------
+
+* [#7](https://github.com/keyko-io/nevermined-gateway/issues/7) Mechanism to encrypt secrets using RSA
+* [#10](https://github.com/keyko-io/nevermined-gateway/issues/10) Access and execution flows improved to check permissions and fulfill
+
+
 0.2.1 (May, 2020)
 -------------------------